Verify methods  and implements valid uuid

+++ server/src/app/controllers/ContactController.js
@@ -1,5 +1,5 @@
 const ContactRepository = require('../repositories/ContactRepository');
-
+const isValidUUIID = require('../utils/IsValidUUID')
 class ContactController {
   async index(request, response) {
     const { orderBy } = request.query;
@@ -10,6 +10,9 @@ class ContactController {
 
   async show(request, response) {
     const { id } = request.params;
+    if(isValidUUIID(id)) {
+      return response.status(400).json({error: 'Invalid contact id'});
+    }
     const contact = await ContactRepository.findById(id);
 
     if (!contact) {
@@ -24,17 +27,22 @@ class ContactController {
       name, email, phone, category_id,
     } = request.body;
 
+    if(category_id && isValidUUIID(category_id)) {
+      return response.status(400).json({error: 'Invalid category id'});
+    }
+
     if (!name) {
       return response.status(400).json({ error: 'Name is required' });
     }
-    const contactExists = await ContactRepository.findByEmail(email);
-
-    if (contactExists) {
-      return response.status(400).json({ error: 'This e-mail is already in use' });
+    if(email) {
+      const contactByEmail = await ContactRepository.findByEmail(email);
+      if (contactByEmail && contactByEmail.id !== id) {
+        return response.status(400).json({ error: 'This e-mail is already in use' });
+      }
     }
 
     const contact = await ContactRepository.create({
-      name, email, phone, category_id,
+      name, email: email || null, phone, category_id: category_id || null ,
     });
 
     response.status(201).json(contact);
@@ -46,22 +54,37 @@ class ContactController {
       name, email, phone, category_id,
     } = request.body;
 
-    const contactExists = await ContactRepository.findById(id);
-    if (!contactExists) {
-      return response.status(404).json({ error: 'User not found' });
+    if(isValidUUIID(id)) {
+      return response.status(400).json({error: 'Invalid contact id'});
+    }
+
+    if(category_id && isValidUUIID(category_id)) {
+      return response.status(400).json({error: 'Invalid category id'});
     }
 
     if (!name) {
       return response.status(400).json({ error: 'Name is required' });
     }
 
-    const contactByEmail = await ContactRepository.findByEmail(email);
-    if (contactByEmail && contactByEmail.id !== id) {
-      return response.status(400).json({ error: 'This e-mail is already in use' });
+    const contactExists = await ContactRepository.findById(id);
+
+    if (!contactExists) {
+      return response.status(404).json({ error: 'Contact not found' });
     }
 
+    if(email) {
+      const contactByEmail = await ContactRepository.findByEmail(email);
+      if (contactByEmail && contactByEmail.id !== id) {
+        return response.status(400).json({ error: 'This e-mail is already in use' });
+      }
+    }
+
+
     const contact = await ContactRepository.update(id, {
-      name, email, phone, category_id,
+      name,
+      email: email || null,
+      phone,
+      category_id:category_id || null,
     });
 
     response.json(contact);
@@ -69,7 +92,9 @@ class ContactController {
 
   async delete(request, response) {
     const { id } = request.params;
-
+    if(isValidUUIID(id)) {
+      return response.status(400).json({error: 'Invalid contact id'});
+    }
     await ContactRepository.delete(id);
 
     response.sendStatus(204);
